+++ src/ui/autocompletes.py
@@ -62,9 +62,10 @@ def bind_autocomplete(callback: Callable, bind_to_object: object) -> Callable:
 @try_except
 async def scanlator(_, current: str) -> list[discord.app_commands.Choice[str]]:
     return [
-        app_commands.Choice(name=_scanlator.name.title(), value=_scanlator.name) for _scanlator in scanlators.values()
-        if _scanlator.name.lower().startswith(current.lower()) and hasattr(_scanlator, "search")
-    ]
+               app_commands.Choice(name=_scanlator.name.title(), value=_scanlator.name) for _scanlator in
+               scanlators.values()
+               if current == "" or _scanlator.name.lower().startswith(current.lower()) and hasattr(_scanlator, "search")
+           ][:25]
 
 
 def get_scanlator_from_current_str(current: str) -> tuple[str | None, str | None]: